Recursive Fill Board
The Swaping of Numbers, Rows, Columns, RowBlocls, ColumnsBlocks, of the Manually Filled Board is Discarded.
The Baord is now Filled by Recursively Placing Numbers in the Boxes.

A Random Number is Selected at First, if its Valid to Put in the Box, then Recusive Fill the Next Box, until All the Boxes are Filled.
If the Number is not Valid, than try Next Number. If All Numbers from 1 to 9, went Invalid, then Revert Back to the Previous Box (Basic Recursion).

Then for the Game, some Numbers are Removed from the Board, with certain Probability.
The Remaining Boxes become Builtin Boxes. Also, the Filled Board is Saved as SolutionBoard

ReadMe File Updated